requirements.txt
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
selenium_wire==5.1.0
